Move components to root folder level
